+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/AzureCommunicationUICallingTests/Redux/Middleware/CallingMiddlewareHandlerTests.swift
@@ -58,8 +58,8 @@ class CallingMiddlewareHandlerTests: XCTestCase {
             expectation.fulfill()
         }
         guard let state: AppState = getState(callingState: .connected,
-                                       cameraStatus: .off,
-                                       cameraDeviceStatus: .front,
+                                             cameraStatus: .off,
+                                             cameraDeviceStatus: .front,
                                              cameraPermission: .notAsked) as? AppState else {
             XCTFail("Failed with state validation")
             return
@@ -105,7 +105,7 @@ class CallingMiddlewareHandlerTests: XCTestCase {
             }
         }
         callingMiddlewareHandler.requestCameraOn(state: getEmptyState(), dispatch: dispatch)
-        wait(for: [expectation], timeout: 1)
+        wait(for: [expectation], timeout: 1.5)
     }
 
     func test_callingMiddlewareHandler_requestCameraOff_when_returnsError_then_updateCameraStatusIsError() {
@@ -129,7 +129,7 @@ class CallingMiddlewareHandlerTests: XCTestCase {
         }
         mockCallingService.error = error
         callingMiddlewareHandler.requestCameraOn(state: getEmptyState(), dispatch: dispatch)
-        wait(for: [expectation], timeout: 1)
+        wait(for: [expectation], timeout: 1.5)
     }
 
     func test_callingMiddlewareHandler_requestCameraSwitch_then_switchCameraCalled() {
@@ -153,7 +153,7 @@ class CallingMiddlewareHandlerTests: XCTestCase {
             }
         }
         callingMiddlewareHandler.requestCameraSwitch(state: getEmptyState(), dispatch: dispatch)
-        wait(for: [expectation], timeout: 1)
+        wait(for: [expectation], timeout: 1.5)
     }
 
     func test_callingMiddlewareHandler_requestCameraSwitch_when_returnsError_then_updateCameraDeviceStatusIsError() {
@@ -165,7 +165,7 @@ class CallingMiddlewareHandlerTests: XCTestCase {
         }
         mockCallingService.error = error
         callingMiddlewareHandler.requestCameraSwitch(state: getEmptyState(), dispatch: dispatch)
-        wait(for: [expectation], timeout: 1)
+        wait(for: [expectation], timeout: 1.5)
     }
 
     func test_callingMiddlewareHandler_endCall_then_endCallCalled() {
@@ -183,14 +183,16 @@ class CallingMiddlewareHandlerTests: XCTestCase {
     func test_callingMiddlewareHandler_endCall_when_returnNSError_then_updateCallError() {
         let expectation = XCTestExpectation(description: "Dispatch the new action")
 
-        let error = getError()
-        let expectedStatus = CallCompositeError(code: CallCompositeErrorCode.callEnd, error: error)
+        let errorCode = 50
+        let error = getError(code: errorCode)
 
         func dispatch(action: Action) {
             XCTAssertTrue(action is ErrorAction.FatalErrorUpdated)
             switch action {
             case let action as ErrorAction.FatalErrorUpdated:
-                XCTAssertEqual(action.error.code, expectedStatus.code)
+                let nserror = action.error as? NSError
+                XCTAssertEqual(nserror!.code, errorCode)
+                XCTAssertEqual(action.internalError, .callEndFailed)
                 expectation.fulfill()
             default:
                 XCTFail("Should not be default \(action)")
@@ -204,13 +206,14 @@ class CallingMiddlewareHandlerTests: XCTestCase {
     func test_callingMiddlewareHandler_endCall_when_returnsCompositeError_then_updateClientError() {
         let expectation = XCTestExpectation(description: "Dispatch the new action")
 
-        let error = CompositeError.invalidSDKWrapper
-        let expectedError = CallCompositeError(code: CallCompositeErrorCode.callEnd, error: error)
+        let error = CallCompositeInternalError.cameraSwitchFailed
+
         func dispatch(action: Action) {
             XCTAssertTrue(action is ErrorAction.FatalErrorUpdated)
             switch action {
             case let action as ErrorAction.FatalErrorUpdated:
-                XCTAssertEqual(action.error.code, expectedError.code)
+                XCTAssertEqual(action.internalError, error)
+                XCTAssertNil(action.error)
                 expectation.fulfill()
             default:
                 XCTFail("Should not be default \(action)")
@@ -222,14 +225,16 @@ class CallingMiddlewareHandlerTests: XCTestCase {
     }
 
     func test_callingMiddlewareHandler_startCall_when_returnsNSError_then_updateCallingCoreError() {
-        let error = getError()
         let expectation = XCTestExpectation(description: "Dispatch the new action")
-        let expectedStatus = CallCompositeError(code: CallCompositeErrorCode.callJoin, error: error)
+        let errorCode = 50
+        let error = getError(code: errorCode)
         func dispatch(action: Action) {
             XCTAssertTrue(action is ErrorAction.FatalErrorUpdated)
             switch action {
             case let action as ErrorAction.FatalErrorUpdated:
-                XCTAssertEqual(action.error.code, expectedStatus.code)
+                let nserror = action.error as? NSError
+                XCTAssertEqual(nserror!.code, errorCode)
+                XCTAssertEqual(action.internalError, .callJoinFailed)
                 expectation.fulfill()
             default:
                 XCTFail("Should not be default \(action)")
@@ -240,16 +245,15 @@ class CallingMiddlewareHandlerTests: XCTestCase {
         wait(for: [expectation], timeout: 1)
     }
 
-    func test_callingMiddlewareHandler_startCall_when_returnsCompositeError_then_updateClientError() {
+    func test_callingMiddlewareHandler_startCall_when_returnsCompositeError_then_updateClientErrorCompositeError() {
         let expectation = XCTestExpectation(description: "Dispatch the new action")
-        let error = CompositeError.invalidSDKWrapper
-        let expectedError = CallCompositeError(code: CallCompositeErrorCode.callJoin, error: error)
+        let error = CallCompositeInternalError.callEndFailed
 
         func dispatch(action: Action) {
             XCTAssertTrue(action is ErrorAction.FatalErrorUpdated)
             switch action {
             case let action as ErrorAction.FatalErrorUpdated:
-                XCTAssertEqual(action.error.code, expectedError.code)
+                XCTAssertEqual(action.internalError, error)
                 expectation.fulfill()
             default:
                 XCTFail("Should not be default \(action)")
@@ -298,14 +302,17 @@ class CallingMiddlewareHandlerTests: XCTestCase {
     }
 
     func test_callingMiddlewareHandler_setupCall_when_returnsError_then_updateCallingCoreError() {
-        let error = getError()
         let expectation = XCTestExpectation(description: "Dispatch the new action")
-        let expectedStatus = CallCompositeError(code: CallCompositeErrorCode.callJoin, error: error)
+        let errorCode = 50
+        let error = getError(code: errorCode)
+
         func dispatch(action: Action) {
             XCTAssertTrue(action is ErrorAction.FatalErrorUpdated)
             switch action {
             case let action as ErrorAction.FatalErrorUpdated:
-                XCTAssertEqual(action.error.code, expectedStatus.code)
+                let nserror = action.error as? NSError
+                XCTAssertEqual(nserror!.code, errorCode)
+                XCTAssertEqual(action.internalError, .callJoinFailed)
                 expectation.fulfill()
             default:
                 XCTFail("Should not be default \(action)")
@@ -313,6 +320,22 @@ class CallingMiddlewareHandlerTests: XCTestCase {
         }
         mockCallingService.error = error
         callingMiddlewareHandler.setupCall(state: getEmptyState(), dispatch: dispatch)
+        wait(for: [expectation], timeout: 1)
+    }
+
+    func test_callingMiddlewareHandler_setupCall_when_internalErrorNotNil_then_shouldNotDispatch() {
+        let expectation = XCTestExpectation(description: "Dispatch the new action")
+        expectation.isInverted = true
+
+        func dispatch(action: Action) {
+            XCTFail("Should not dispatch")
+        }
+        callingMiddlewareHandler.setupCall(state: getState(callingState: .none,
+                                                           cameraStatus: .off,
+                                                           cameraDeviceStatus: .front,
+                                                           cameraPermission: .granted,
+                                                           internalError: .callEvicted),
+                                           dispatch: dispatch)
         wait(for: [expectation], timeout: 1)
     }
 
@@ -404,7 +427,31 @@ class CallingMiddlewareHandlerTests: XCTestCase {
                                                                  cameraStatus: .paused,
                                                                  cameraDeviceStatus: .front),
                                                  dispatch: dispatch)
-        wait(for: [expectation], timeout: 1)
+        wait(for: [expectation], timeout: 1.5)
+    }
+
+    func test_callingMiddlewareHandler_holdCall_then_holdCallCalled() {
+        guard let state: AppState = getState(callingState: .connected,
+                                             cameraStatus: .off,
+                                             cameraDeviceStatus: .front,
+                                             cameraPermission: .notAsked) as? AppState else {
+            XCTFail("Failed with state validation")
+            return
+        }
+        callingMiddlewareHandler.holdCall(state: state, dispatch: getEmptyDispatch())
+        XCTAssertTrue(mockCallingService.holdCallCalled)
+    }
+
+    func test_callingMiddlewareHandler_resumeCall_then_resumeCallCalled() {
+        guard let state: AppState = getState(callingState: .localHold,
+                                             cameraStatus: .off,
+                                             cameraDeviceStatus: .front,
+                                             cameraPermission: .notAsked) as? AppState else {
+            XCTFail("Failed with state validation")
+            return
+        }
+        callingMiddlewareHandler.resumeCall(state: state, dispatch: getEmptyDispatch())
+        XCTAssertTrue(mockCallingService.resumeCallCalled)
     }
 
     func test_callingMiddlewareHandler_enterForeground_when_callLocalHold_cameraStatusOn_noError_then_updateCameraStatusOnUpdate() {
@@ -425,7 +472,7 @@ class CallingMiddlewareHandlerTests: XCTestCase {
                                                                  cameraStatus: .paused,
                                                                  cameraDeviceStatus: .front),
                                                  dispatch: dispatch)
-        wait(for: [expectation], timeout: 1)
+        wait(for: [expectation], timeout: 1.5)
     }
 
     func test_callingMiddlewareHandler_enterForeground_when_callConnected_cameraStatusOn_returnsError_then_updateCameraStatusIsError() {
@@ -446,7 +493,7 @@ class CallingMiddlewareHandlerTests: XCTestCase {
         }
         callingMiddlewareHandler.onCameraPermissionIsSet(state: getState(cameraPermission: .requesting,
                                                                          cameraTransmissionStatus: .local),
-                                                 dispatch: dispatch)
+                                                         dispatch: dispatch)
     }
 
     func test_callingMiddlewareHandler_onCameraPermissionIsSet_when_callTransmissionRemote_cameraPermissionRequesting_then_updateCameraOnTriggered() {
@@ -455,7 +502,7 @@ class CallingMiddlewareHandlerTests: XCTestCase {
         }
         callingMiddlewareHandler.onCameraPermissionIsSet(state: getState(cameraPermission: .requesting,
                                                                          cameraTransmissionStatus: .remote),
-                                                 dispatch: dispatch)
+                                                         dispatch: dispatch)
     }
 
     func test_callingMiddlewareHandler_onCameraPermissionIsSet_when_callTransmissionRemote_cameraPermissionNotRequesting_then_updateCameraOnTriggered() {
@@ -464,7 +511,7 @@ class CallingMiddlewareHandlerTests: XCTestCase {
         }
         callingMiddlewareHandler.onCameraPermissionIsSet(state: getState(cameraPermission: .granted,
                                                                          cameraTransmissionStatus: .remote),
-                                                 dispatch: dispatch)
+                                                         dispatch: dispatch)
     }
 }
 
@@ -478,7 +525,8 @@ extension CallingMiddlewareHandlerTests {
                           cameraStatus: LocalUserState.CameraOperationalStatus = .on,
                           cameraDeviceStatus: LocalUserState.CameraDeviceSelectionStatus = .front,
                           cameraPermission: AppPermission.Status = .unknown,
-                          cameraTransmissionStatus: LocalUserState.CameraTransmissionStatus = .local) -> ReduxState {
+                          cameraTransmissionStatus: LocalUserState.CameraTransmissionStatus = .local,
+                          internalError: CallCompositeInternalError? = nil) -> ReduxState {
         let callState = CallingState(status: callingState)
         let cameraState = LocalUserState.CameraState(operation: cameraStatus,
                                                      device: cameraDeviceStatus,
@@ -491,19 +539,21 @@ extension CallingMiddlewareHandlerTests {
                                         localVideoStreamIdentifier: nil)
         let permissionState = PermissionState(audioPermission: .unknown,
                                               cameraPermission: cameraPermission)
+        let errorState = ErrorState(internalError: internalError)
         return AppState(callingState: callState,
                         permissionState: permissionState,
                         localUserState: localState,
                         lifeCycleState: LifeCycleState(),
-                        remoteParticipantsState: .init())
+                        remoteParticipantsState: .init(),
+                        errorState: errorState)
     }
 
     private func getEmptyDispatch() -> ActionDispatch {
         return { _ in }
     }
 
-    private func getError() -> Error {
-        return NSError(domain: "", code: 100, userInfo: [
+    private func getError(code: Int = 100) -> Error {
+        return NSError(domain: "", code: code, userInfo: [
             NSLocalizedDescriptionKey: "Error"
         ])
     }